SetTest, StringTest 코드 리뷰사항 적용
